refactor: separation of concern
